+++ Source/data_transfer/transferer.py
@@ -1,12 +1,10 @@
 # Standard library imports
 import dataclasses
-import itertools
 import queue
 import uuid
 
 # Typing imports
 from typing import Any
-
 
 
 @dataclasses.dataclass
@@ -19,18 +17,23 @@ class _input_type:
 class obj_type:
     def __init__(self):
         super().__init__()
+        self.was_triggered = False
         self.input_queue = queue.Queue[_input_type]()
         self.output_dict = dict[str, queue.Queue[_input_type]]()
 
-    def generate_guid(self) -> str:
+    def _generate_guid(self) -> str:
         while True:
             guid = str(uuid.uuid4())
             if guid not in self.output_dict:
                 return guid
 
     def call(self, path: str, game_config, *call_args):
+        '''
+        This function is accessed by Python scripts which want to interact with RCC servers.
+        The function signature should match whatever `gen_function` in `callable.py` specified (as of 2026-06-05).
+        '''
         temp_queue = queue.Queue()
-        guid = self.generate_guid()
+        guid = self._generate_guid()
         self.output_dict[guid] = temp_queue
 
         self.input_queue.put(_input_type(
@@ -45,6 +48,10 @@ class obj_type:
         return result
 
     def extract(self) -> dict[str, dict[str, Any]]:
+        '''
+        This function is accessed by the webserver.
+        '''
+        self.was_triggered = True
         result = {}
         try:
             item = self.input_queue.get(