+++ Source/game_config/structure.py
@@ -13,6 +13,7 @@ import assets
 class config_type(allocateable.obj_type):
     '''
     Configuration specification, according by default to "GameConfig.toml".
+    More info for each type parameter should be found in Rōblox Freedom Distribution's main README.md file.
     '''
     class metadata(allocateable.obj_type):
         config_version_wildcard: wrappers.rfd_version_check = "*"
@@ -36,6 +37,7 @@ class config_type(allocateable.obj_type):
         # Don't count too much on 2021E.
         # I really recommend that people manually specify which version of Rōblox they want to run.
         roblox_version: util.versions.rōblox = util.versions.rōblox.v463
+        ready_delay_sec: float = 3
 
     class server_core(allocateable.obj_type):
         class place_file(allocateable.obj_type):