+++ Source/routines/rcc/log_action.py
@@ -5,6 +5,7 @@ class LogAction(enum.Enum):
     PROCEED = 0
     RESTART = 1
     TERMINATE = 2
+    READY = 3
 
 
 def check(line: bytes) -> LogAction:
@@ -12,4 +13,6 @@ def check(line: bytes) -> LogAction:
         return LogAction.RESTART
     elif line.endswith(b'crashes to Backtrace'):
         return LogAction.RESTART
+    elif b'Finished initializing game' in line:
+        return LogAction.READY
     return LogAction.PROCEED