+++ Source/routines/rcc/startup_scripts.py
@@ -3,6 +3,8 @@ import game_config
 
 BASE_SCRIPT_FORMAT = """\
 %(rcc_snippet)s
+
+wait(%(ready_delay_sec)f)
 local BaseUrl = game:GetService("ContentProvider").BaseUrl:lower()
 local HttpRbxApiService = game:GetService("HttpRbxApiService")
 local HttpService = game:GetService("HttpService")
@@ -44,6 +46,7 @@ print('Initialised RFD server scripts.')
 
 def get_script(game_config: game_config.obj_type) -> str:
     return BASE_SCRIPT_FORMAT % {
+        'ready_delay_sec': game_config.game_setup.ready_delay_sec,
         'rcc_snippet': data_transfer.get_rcc_snippet(game_config),
         'startup_script': game_config.server_core.startup_script,
     }