+++ Source/routines/web.py
@@ -28,7 +28,7 @@ class obj_type(logic.gameconfig_entry, logic.loggable_entry):
     @override
     def process(self) -> None:
         super().process()
-        self.httpd = web_server.make_server(
+        self.httpd = web_server.make_httpd(
             self.web_port,
             self.is_ssl,
             self.is_ipv6,